+++ config/babel.prod.js
@@ -8,22 +8,34 @@
  */
 
 module.exports = {
+  // Don't try to find .babelrc because we want to force this configuration.
   babelrc: false,
   presets: [
-    'babel-preset-es2015',
-    'babel-preset-es2016',
-    'babel-preset-react'
-  ].map(require.resolve),
+    // let, const, destructuring, classes, modules
+    require.resolve('babel-preset-es2015'),
+    // exponentiation
+    require.resolve('babel-preset-es2016'),
+    // JSX, Flow
+    require.resolve('babel-preset-react')
+  ],
   plugins: [
-    'babel-plugin-syntax-trailing-function-commas',
-    'babel-plugin-transform-class-properties',
-    'babel-plugin-transform-object-rest-spread',
-    'babel-plugin-transform-react-constant-elements',
-  ].map(require.resolve).concat([
+    // function x(a, b, c,) { }
+    require.resolve('babel-plugin-syntax-trailing-function-commas'),
+    // await fetch()
+    require.resolve('babel-plugin-syntax-async-functions'),
+    // class { handleClick = () => { } }
+    require.resolve('babel-plugin-transform-class-properties'),
+    // { ...todo, completed: true }
+    require.resolve('babel-plugin-transform-object-rest-spread'),
+    // function* () { yield 42; yield 43; }
+    require.resolve('babel-plugin-transform-regenerator'),
+    // Polyfills the runtime needed for async/await and generators
     [require.resolve('babel-plugin-transform-runtime'), {
       helpers: false,
       polyfill: false,
       regenerator: true
-    }]
-  ])
+    }],
+    // Optimization: hoist JSX that never changes out of render()
+    require.resolve('babel-plugin-transform-react-constant-elements')
+  ]
 };