+++ config/polyfills.js
@@ -6,4 +6,5 @@ if (typeof Promise === 'undefined') {
   window.Promise = require('promise/lib/es6-extensions.js');
 }
 
+// fetch() polyfill for making API calls.
 require('whatwg-fetch');