+++ scripts/init.js
@@ -19,13 +19,17 @@ module.exports = function(appPath, appName, verbose, originalDirectory) {
 
   // Copy over some of the devDependencies
   appPackage.dependencies = appPackage.dependencies || {};
+  appPackage.devDependencies = appPackage.devDependencies || {};
   ['react', 'react-dom'].forEach(function (key) {
     appPackage.dependencies[key] = ownPackage.devDependencies[key];
+  });
+  ['react-test-renderer'].forEach(function (key) {
+    appPackage.devDependencies[key] = ownPackage.devDependencies[key];
   });
 
   // Setup the script rules
   appPackage.scripts = {};
-  ['start', 'build', 'eject'].forEach(function(command) {
+  ['start', 'build', 'eject', 'test'].forEach(function(command) {
     appPackage.scripts[command] = 'react-scripts ' + command;
   });